flask/requirements.txt
Flask
uwsgi
gensim
numpy
scipy
mecab-python3
pyyaml
